+++ crates/client/data-availability/src/avail/config.rs
@@ -1,6 +1,3 @@
-use std::fs::File;
-use std::path::PathBuf;
-
 use serde::Deserialize;
 
 use crate::{DaError, DaLayer, DaMode};