+++ crates/client/data-availability/src/ethereum/config.rs
@@ -1,6 +1,3 @@
-use std::fs::File;
-use std::path::PathBuf;
-
 use serde::{Deserialize, Serialize};
 
 use crate::{DaError, DaMode};